pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coda"
version = "2.0.0-alpha"
description = "Core Operations & Digital Assistant - Next-generation local-first voice assistant"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Coda Development Team"},
]
keywords = ["voice-assistant", "ai", "local-first", "real-time", "speech"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
requires-python = ">=3.10"

dependencies = [
    # Core dependencies
    "pydantic>=2.11.7,<3.0.0",  # Pin before V3 breaking changes
    "pyyaml>=6.0.1",
    "python-dotenv>=1.1.0",
    "click>=8.1.7",

    # Async and networking (latest versions)
    "websockets>=15.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",

    # Audio processing (latest)
    "numpy>=2.0.0",
    "soundfile>=0.13.0",
    "librosa>=0.11.0",
    "pyaudio>=0.2.11",

    # Machine Learning (RTX 5090 optimized)
    "torch>=2.6.0",  # Will install nightly for SM_120
    "transformers>=4.53.0",
    "sentence-transformers>=5.0.0",

    # Memory and storage (latest)
    "chromadb>=1.0.10",

    # LLM integration
    "ollama>=0.5.1",
    "httpx>=0.28.0",  # Latest async HTTP
    "requests>=2.32.0",

    # Utilities (latest)
    "tqdm>=4.67.0",
    "rich>=14.0.0",
    "structlog>=24.0.0",

    # Performance and monitoring
    "psutil>=7.0.0",
    "pynvml>=12.0.0",
]

[project.optional-dependencies]
# Voice processing dependencies
voice = [
    "moshi>=0.2.9",
    "webrtcvad>=2.0.10",
    "silero-vad>=5.0.0",
    "noisereduce>=3.1.0",
    "aiofiles>=24.1.0",
]

# Performance optimization dependencies
performance = [
    "bitsandbytes>=0.45.0",
    "optimum>=1.20.0",
    "accelerate>=1.0.0",
    "diskcache>=5.6.3",
    "redis>=5.0.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

# Production deployment
prod = [
    "gunicorn>=21.0.0",
    "docker>=6.0.0",
]

# GPU acceleration
gpu = [
    "torch[cuda]>=2.0.0",
    "torchaudio>=2.0.0",
    "accelerate>=0.20.0",
]

# All optional dependencies
all = [
    "coda[kyutai,dev,prod,gpu]"
]

[project.urls]
Homepage = "https://github.com/yourusername/coda"
Documentation = "https://github.com/yourusername/coda/docs"
Repository = "https://github.com/yourusername/coda"
Issues = "https://github.com/yourusername/coda/issues"

[project.scripts]
coda = "coda.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/coda"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "/configs",
    "/scripts",
]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["coda"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pyaudio.*",
    "soundfile.*",
    "chromadb.*",
    "ollama.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=src/coda",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests that require GPU",
    "asyncio: marks tests as async tests",
    "performance: marks tests as performance benchmarks",
    "load_test: marks tests as load tests",
    "stress_test: marks tests as stress tests",
    "recovery_test: marks tests as recovery tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/coda"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
